% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestNoaaGauge.R
\name{findNearestNoaaGauge}
\alias{findNearestNoaaGauge}
\title{Find the nearest NOAA gauge to each site}
\usage{
findNearestNoaaGauge(sites, site_lat = "latitude", site_lon = "longitude")
}
\arguments{
\item{sites}{A data frame, tibble, or \code{sf} object containing site locations.}

\item{site_lat}{Character. Name of the latitude column in \code{sites}.
Default is \code{"latitude"}.}

\item{site_lon}{Character. Name of the longitude column in \code{sites}.
Default is \code{"longitude"}.}
}
\value{
An data.frame object containing the original site attributes and geometry,
attributes of the nearest gauge, and distance to the gauge in meters and
kilometers.
}
\description{
Finds the nearest long term NOAA gauge to each input site using
\code{sf::st_nearest_feature()}, joins the corresponding gauge attributes to
the site table, and calculates the straight-line distance between the site
and gauge.
}
\details{
\code{sites} may be data frame containing latitude and
longitude columns or existing \code{sf} objects.

Distances are straight-line geographic distances. For coastal and estuarine
applications, the nearest gauge by straight-line distance may not
necessarily be the most hydrologically connected gauge. This also won't work outside the U.S.
}
