% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFullTidalScenario.R
\name{generateFullTidalScenario}
\alias{generateFullTidalScenario}
\title{Function to query NOAA tide gauge information and generate annual flood level scenearios}
\usage{
generateFullTidalScenario(
  station_id = 9410660,
  run_hindcast = T,
  run_forecast = T,
  hindcast_start = 1923,
  forecast_start = 2018,
  forecast_end = 2100,
  RCP = c("RCP4.5"),
  RCP_probability = c(0.25, 0.5, 0.75),
  include_lt_tidal_const = T,
  datum_start_year = 1980,
  datum_end_year = 2025
)
}
\arguments{
\item{station_id}{Unique NOAA station identifier}

\item{run_hindcast}{True or False, run a hindcast}

\item{run_forecast}{True or False, run a forecast}

\item{hindcast_start}{Calendar year (YYYY) that hindcast starts}

\item{forecast_start}{Calendar year (YYYY) that forecast starts}

\item{forecast_end}{Calendar year (YYYY) that forecast ends, needs to be 2030, 2050, 2100, 2150 or 2200}

\item{RCP}{Realized concentration pathway, either string or vector, must be RCP2.6, RCP4.5 or RCP8.5}

\item{RCP_probability}{RCP probability, either numeric or vector, sea-level rise senario probabilities to test.}

\item{include_lt_tidal_const}{Include long term tidal constituents, True or False.}

\item{datum_start_year}{Datum start year over which to calculate tidal datums (calendar year, YYYY).}

\item{datum_end_year}{Datum end year over which to calculate tidal datums (calendar year, YYYY).}
}
\description{
Generate a full sea-level rise and tidal inundation scenario
}
\author{
J. Holmquist
}
